zAxis functionality works, some skewed positioning before click on carousel
with > 3 facets, carousel moves to side when adjusting z axis, but upon
click it re-centers with the correct changed z axis. If no z axis adjustments
are made, everything moves/changes as expected. The minor deviation before
clicking on a facet when adjusting the z axis is acceptable. This utility is
meant for previewing various positioning effects.